# chore: upgrade @typescript-eslint to v6 and eslint to v8