--- a/change/@fluentui-eslint-plugin-912d0763-d6cf-4fec-beac-2843c49fc62d.json
+++ b/change/@fluentui-eslint-plugin-912d0763-d6cf-4fec-beac-2843c49fc62d.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "chore: migrate to @typescript/eslint v6",
+  "packageName": "@fluentui/eslint-plugin",
+  "email": "vgenaev@gmail.com",
+  "dependentChangeType": "patch"
+}