--- a/change/@fluentui-react-tabster-b0807315-3821-4fdb-a2ac-25c10dd1d266.json
+++ b/change/@fluentui-react-tabster-b0807315-3821-4fdb-a2ac-25c10dd1d266.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: suppress @typescript-eslint/naming-convention report on TabsterTypes_6_0_1_DoNotUse",
+  "packageName": "@fluentui/react-tabster",
+  "email": "vgenaev@gmail.com",
+  "dependentChangeType": "none"
+}