--- a/change/@fluentui-web-components-8db325bc-acef-4d3a-af10-591cfd00cc42.json
+++ b/change/@fluentui-web-components-8db325bc-acef-4d3a-af10-591cfd00cc42.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: suppress @typescript-eslint/no_unsafe_declaration-merging reports",
+  "packageName": "@fluentui/web-components",
+  "email": "vgenaev@gmail.com",
+  "dependentChangeType": "none"
+}