--- a/packages/eslint-plugin/package.json
+++ b/packages/eslint-plugin/package.json
@@ -15,27 +15,29 @@
   },
   "dependencies": {
     "@griffel/eslint-plugin": "^1.6.3",
-    "@rnx-kit/eslint-plugin": "^0.4.2",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/utils": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
+    "@rnx-kit/eslint-plugin": "^0.5.3",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/utils": "^6.21.0",
+    "@typescript-eslint/rule-tester": "6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
+    "@typescript-eslint/type-utils": "^6.21.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-typescript": "^2.7.1",
-    "eslint-plugin-deprecation": "^1.2.1",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-deprecation": "^2.0.0",
     "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^26.9.0",
-    "eslint-plugin-jsdoc": "^48.2.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-jest": "^27.9.0",
+    "eslint-plugin-jsdoc": "^48.7.0",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
     "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "fs-extra": "^8.1.0",
     "minimatch": "^3.1.2",
     "jju": "^1.4.0"
   },
   "peerDependencies": {
-    "eslint": "^7.0.0",
-    "typescript": "^4.1.0"
+    "eslint": "^8.0.0",
+    "typescript": "^4.2.4"
   },
   "files": [
     "src"