--- a/packages/eslint-plugin/src/rules/ban-imports.js
+++ b/packages/eslint-plugin/src/rules/ban-imports.js
@@ -31,7 +31,6 @@ module.exports = createRule({
     type: 'problem',
     docs: {
       description: 'Ban importing (or re-exporting) certain identifiers from certain paths or modules.',
-      recommended: false,
     },
     messages: {
       pathNotAllowed: "{{verb}} from '{{path}}' is not allowed{{message}}",