--- a/packages/eslint-plugin/src/rules/max-len.js
+++ b/packages/eslint-plugin/src/rules/max-len.js
@@ -10,7 +10,6 @@ module.exports = createRule({
       // - it has options to detect and specially handle comments, strings, etc
       // - it checks the ignore regex without even a preliminary length check first
       description: 'Enforces a maximum line length, more cheaply than default ESLint version',
-      recommended: false,
     },
     messages: {
       max: 'This line has a length of {{lineLength}}. Maximum allowed is {{max}}.',