--- a/packages/eslint-plugin/src/rules/no-tslint-comments.js
+++ b/packages/eslint-plugin/src/rules/no-tslint-comments.js
@@ -7,7 +7,7 @@ module.exports = createRule({
     type: 'problem',
     docs: {
       description: 'Forbid tslint:disable/tslint:enable comments after ESLint migration.',
-      recommended: 'error',
+      recommended: 'recommended',
     },
     messages: {
       tslint: 'tslint:{{verb}} comments are unnecessary with ESLint',