--- a/packages/react-components/react-tabster/src/index.ts
+++ b/packages/react-components/react-tabster/src/index.ts
@@ -57,6 +57,7 @@ export { KEYBORG_FOCUSIN } from 'keyborg';
 
 // WARNING! ATTENTION! Tabster.Types was exported from here by mistake. To avoid breaking changes,
 // we are putting a snapshot of Tabster.Types@6.0.1 and marking the entire export as deprecated.
+// eslint-disable-next-line @typescript-eslint/naming-convention
 import * as TabsterTypes6_0_1_DoNotUse from './tabster-types-6.0.1-do-not-use';
 export {
   /** @deprecated (Do not use! Exposed by mistake and will be removed in the next major version.)  */