--- a/scripts/lint-staged/src/eslint-for-package.js
+++ b/scripts/lint-staged/src/eslint-for-package.js
@@ -1,7 +1,7 @@
 // @ts-check
 
-const fs = require('fs');
-const path = require('path');
+const fs = require('node:fs');
+const path = require('node:path');
 
 const { eslintConstants } = require('@fluentui/scripts-monorepo');
 const { ESLint } = require('eslint');