--- a/tools/workspace-plugin/src/generators/eslint-rule/generator.spec.ts
+++ b/tools/workspace-plugin/src/generators/eslint-rule/generator.spec.ts
@@ -68,12 +68,10 @@ describe('eslint-rule generator', () => {
     `);
 
     expect(tree.read(paths.spec, 'utf-8')).toMatchInlineSnapshot(`
-      "import { TSESLint } from '@typescript-eslint/utils';
+      "import { RuleTester } from '@typescript-eslint/rule-tester';
       import { rule, RULE_NAME } from './uppercase';
 
-      const ruleTester = new TSESLint.RuleTester({
-        parser: require.resolve('@typescript-eslint/parser'),
-      });
+      const ruleTester = new RuleTester();
 
       ruleTester.run(RULE_NAME, rule, {
         valid: [\`const example = true;\`],