--- a/tools/workspace-plugin/src/generators/eslint-rule/generator.spec.ts
+++ b/tools/workspace-plugin/src/generators/eslint-rule/generator.spec.ts
@@ -70,9 +70,7 @@ describe('eslint-rule generator', () => {
     expect(tree.read(paths.spec, 'utf-8')).toMatchInlineSnapshot(`
       "import { RuleTester } from '@typescript-eslint/rule-tester';
       import { rule, RULE_NAME } from './uppercase';
-
       const ruleTester = new RuleTester();
-
       ruleTester.run(RULE_NAME, rule, {
         valid: [\`const example = true;\`],
         invalid: [],