<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

<!-- This is the behavior we have today -->

## New Behavior

- Upgraded `@typescript-eslint` to v6 and eslint to v8.
- Updated related eslint packages
- Added `@typescript-eslint/rule-tester` package

**api changes**:
- `meta.docs.recommended` now `recommended | stylistic | error` instead of `error | strict | warn | false`
- `@typescript-eslint/utils` no longer exports `RuleTester` 
- `CLIEngine` removed

**Perf bench**

`yarn lage lint`

| Before | After  |
|--------|--------|
| 10m 6s | 10m 20s / **𝚫 +2.3% slowdown** | 

## Related Issue(s)

- Partially implements #30334