@fabricteam 2024-07-12
## [Perf Analysis (`@fluentui/react-components`)](https://github.com/microsoft/fluentui/wiki/Perf-Testing)
<p>No significant results to display.</p><details><summary>All results</summary><p>
  <table>
  <tr>
    <th>Scenario</th>
    <th>Render type</th>
    <th>
      <a href="https://github.com/microsoft/fluentui/wiki/Perf-Testing#why-are-results-listed-in-ticks-instead-of-time-units">Master Ticks</a>
    </th>
    <th>
      <a href="https://github.com/microsoft/fluentui/wiki/Perf-Testing#why-are-results-listed-in-ticks-instead-of-time-units">PR Ticks</a>
    </th>
    <th>Iterations</th>
    <th>Status</th>
  </tr><tr>
            <td>Avatar</td>
            <td>mount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/Avatar-mount_base.html">653</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/Avatar-mount.html">670</a></td>
            <td>5000</td>
            <td></td>
           </tr>
<tr>
            <td>Button</td>
            <td>mount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/Button-mount_base.html">312</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/Button-mount.html">286</a></td>
            <td>5000</td>
            <td></td>
           </tr>
<tr>
            <td>Field</td>
            <td>mount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/Field-mount_base.html">1153</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/Field-mount.html">1114</a></td>
            <td>5000</td>
            <td></td>
           </tr>
<tr>
            <td>FluentProvider</td>
            <td>mount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/FluentProvider-mount_base.html">733</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/FluentProvider-mount.html">718</a></td>
            <td>5000</td>
            <td></td>
           </tr>
<tr>
            <td>FluentProviderWithTheme</td>
            <td>mount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/FluentProviderWithTheme-mount_base.html">89</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/FluentProviderWithTheme-mount.html">91</a></td>
            <td>10</td>
            <td></td>
           </tr>
<tr>
            <td>FluentProviderWithTheme</td>
            <td>virtual-rerender</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/FluentProviderWithTheme-virtual-rerender_base.html">41</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/FluentProviderWithTheme-virtual-rerender.html">36</a></td>
            <td>10</td>
            <td></td>
           </tr>
<tr>
            <td>FluentProviderWithTheme</td>
            <td>virtual-rerender-with-unmount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/FluentProviderWithTheme-virtual-rerender-with-unmount_base.html">79</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/FluentProviderWithTheme-virtual-rerender-with-unmount.html">88</a></td>
            <td>10</td>
            <td></td>
           </tr>
<tr>
            <td>MakeStyles</td>
            <td>mount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/MakeStyles-mount_base.html">892</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/MakeStyles-mount.html">876</a></td>
            <td>50000</td>
            <td></td>
           </tr>
<tr>
            <td>Persona</td>
            <td>mount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/Persona-mount_base.html">1781</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/Persona-mount.html">1717</a></td>
            <td>5000</td>
            <td></td>
           </tr>
<tr>
            <td>SpinButton</td>
            <td>mount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/SpinButton-mount_base.html">1397</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/SpinButton-mount.html">1372</a></td>
            <td>5000</td>
            <td></td>
           </tr>
<tr>
            <td>SwatchPicker</td>
            <td>mount</td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/SwatchPicker-mount_base.html">1601</a></td>
            <td><a href="https://fluentuipr.z22.web.core.windows.net/pull/31988/perf-test-react-components/SwatchPicker-mount.html">1680</a></td>
            <td>5000</td>
            <td></td>
           </tr></table></p></details>

<div id="perfComment9425"/>